+++ frontend/src/app/duplicate-alert.component.spec.ts
@@ -1,6 +1,4 @@
-import { Component } from '@angular/core';
 import { ComponentFixture, TestBed, async } from '@angular/core/testing';
-import { RouterTestingModule } from '@angular/router/testing';
 import { StoreModule } from '@ngrx/store';
 
 import { Router } from '@angular/router';
@@ -13,47 +11,34 @@ import * as WebSocketActions from './state/actions/websocket';
 import { State } from './state/reducer';
 import { websocketReducer } from './state/reducers/websocket';
 
-import { AppComponent } from './app.component';
+import { DuplicateAlertComponent } from './duplicate-alert.component';
 
-let comp: AppComponent;
-let fixture: ComponentFixture<AppComponent>;
+import { filterCallByAction } from './testing';
 
-@Component({
-  selector: 'app-menu-bar',
-  template: ''
-})
-class MockMenuBarComponent {}
 
-@Component({
-  selector: 'app-duplicate-alert',
-  template: ''
-})
-class MockDuplicateAlertComponent {}
-
-describe('AppComponent', () => {
+describe('DuplicateAlertComponent', () => {
+  let comp: DuplicateAlertComponent;
+  let fixture: ComponentFixture<DuplicateAlertComponent>;
   let store: Store<State>;
+  let dispatchSpy;
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
       declarations: [
-        AppComponent,
-        MockMenuBarComponent,
-        MockDuplicateAlertComponent,
+        DuplicateAlertComponent,
       ],
       imports: [
-        RouterTestingModule.withRoutes([]),
         StoreModule.forRoot({
           websocket: websocketReducer,
         }),
       ],
-      providers: [
-      ],
     }).compileComponents()
     .then(() => {
-      fixture = TestBed.createComponent(AppComponent);
+      fixture = TestBed.createComponent(DuplicateAlertComponent);
       comp = fixture.componentInstance;
 
       store = fixture.debugElement.injector.get(Store);
+      dispatchSpy = spyOn(store, 'dispatch').and.callThrough();
     });
   }));
 
@@ -62,14 +47,17 @@ describe('AppComponent', () => {
   });
 
   it(
-    'should display DuplicateAlertComponent when connectionStatus is duplicate',
+    'should force connect when Force Connect button is clicked',
     async(() => {
-      store.dispatch(new WebSocketActions.DuplicateSession());
+      const button = fixture.nativeElement.querySelector('.force-connect-button');
+      button.click();
       fixture.detectChanges();
       fixture.whenStable().then(() => {
         fixture.detectChanges();
-        const dupAlert = fixture.nativeElement.querySelector('app-duplicate-alert');
-        expect(dupAlert).not.toBeNull();
+        const connect = filterCallByAction(dispatchSpy, WebSocketActions.Connect);
+        expect(connect).toEqual([
+          new WebSocketActions.Connect(true),
+        ]);
       });
     })
   );
